+++ elasticai/creator/tests/vhdl/components/test_fp_linear_1d_component.py
@@ -9,11 +9,11 @@ class FPLinear1dComponentTest(unittest.TestCase):
         to_fp = FixedPoint.get_factory(total_bits=8, frac_bits=4)
 
         component = FPLinear1dComponent(
-            layer_name="ll1",
+            layer_id="ll1",
             in_features=3,
             out_features=2,
             fixed_point_factory=to_fp,
             work_library_name="work",
         )
 
-        self.assertEqual(len(list(component())), 204)
+        self.assertEqual(len(list(component())), 213)