+++ elasticai/creator/tests/vhdl/translator/abstract/layers/test_fp_linear_1d_module.py
@@ -11,7 +11,7 @@ from elasticai.creator.vhdl.translator.abstract.layers import (
 class FPLinear1dModuleTest(unittest.TestCase):
     def setUp(self) -> None:
         self.linear = FPLinear1dModule(
-            layer_name="ll1", weight=[[1.0, 2.0, 3.0, 4.0, 5.0, 6.0]], bias=[1.0, 2.0]
+            layer_id="ll1", weight=[[1.0, 2.0, 3.0, 4.0, 5.0, 6.0]], bias=[1.0, 2.0]
         )
         self.translation_args = FPLinear1dTranslationArgs(
             fixed_point_factory=FixedPoint.get_factory(total_bits=8, frac_bits=4)
@@ -23,20 +23,18 @@ class FPLinear1dModuleTest(unittest.TestCase):
         target_components = [
             (
                 FPLinear1dComponent,
-                "fp_linear_1d_{layer_name}.vhd".format(
-                    layer_name=self.linear.layer_name
-                ),
+                "fp_linear_1d_{layer_name}.vhd".format(layer_name=self.linear.layer_id),
             ),
             (
                 RomComponent,
                 "w_rom_fp_linear_1d_{layer_name}.vhd".format(
-                    layer_name=self.linear.layer_name
+                    layer_name=self.linear.layer_id
                 ),
             ),
             (
                 RomComponent,
                 "b_rom_fp_linear_1d_{layer_name}.vhd".format(
-                    layer_name=self.linear.layer_name
+                    layer_name=self.linear.layer_id
                 ),
             ),
         ]