+++ elasticai/creator/vhdl/components/fp_hard_sigmoid_componet.py
@@ -7,10 +7,12 @@ from elasticai.creator.vhdl.number_representations import (
     FixedPoint,
     fixed_point_params_from_factory,
 )
+from elasticai.creator.vhdl.templates.utils import expand_template
 
 
 @dataclass
 class FPHardSigmoidComponent:
+    layer_id: str  # used to distinguish layers in the same model
     zero_threshold: FixedPoint
     one_threshold: FixedPoint
     slope: FixedPoint
@@ -24,14 +26,16 @@ class FPHardSigmoidComponent:
 
     @property
     def file_name(self) -> str:
-        return "fp_hard_sigmoid.vhd"
+        return f"fp_hard_sigmoid_{self.layer_id}.vhd"
 
     def __call__(self) -> Code:
         template = read_text(
             "elasticai.creator.vhdl.templates", "fp_hard_sigmoid.tpl.vhd"
         )
 
-        code = template.format(
+        code = expand_template(
+            template.splitlines(),
+            layer_name=self.layer_id,
             data_width=self.data_width,
             frac_width=self.frac_width,
             one=self.fixed_point_factory(1).to_signed_int(),
@@ -41,4 +45,4 @@ class FPHardSigmoidComponent:
             y_intercept=self.y_intercept.to_signed_int(),
         )
 
-        yield from code.splitlines()
+        yield from code