+++ elasticai/creator/vhdl/components/fp_linear_1d_component.py
@@ -8,11 +8,12 @@ from elasticai.creator.vhdl.components.utils import (
 )
 from elasticai.creator.vhdl.language import Code
 from elasticai.creator.vhdl.number_representations import FixedPoint
+from elasticai.creator.vhdl.templates.utils import expand_template
 
 
 @dataclass
 class FPLinear1dComponent:
-    layer_name: str  # used to distinguish layers in the same model
+    layer_id: str  # used to distinguish layers in the same model
     in_features: int
     out_features: int
     fixed_point_factory: Callable[[float], FixedPoint]
@@ -28,13 +29,14 @@ class FPLinear1dComponent:
 
     @property
     def file_name(self) -> str:
-        return "fp_linear_1d_{layer_name}.vhd".format(layer_name=self.layer_name)
+        return f"fp_linear_1d_{self.layer_id}.vhd"
 
     def __call__(self) -> Code:
         template = read_text("elasticai.creator.vhdl.templates", "fp_linear_1d.tpl.vhd")
 
-        code = template.format(
-            layer_name=self.layer_name,
+        code = expand_template(
+            template.splitlines(),
+            layer_name=self.layer_id,
             work_library_name=self.work_library_name,
             data_width=self.data_width,
             frac_width=self.frac_width,
@@ -45,4 +47,4 @@ class FPLinear1dComponent:
             resource_option=f'"{self.resource_option}"',
         )
 
-        yield from code.splitlines()
+        yield from code