+++ elasticai/creator/vhdl/translator/pytorch/build_functions/fp_relu_build_function.py
@@ -4,5 +4,7 @@ from elasticai.creator.vhdl.translator.abstract.layers.fp_relu_module import (
 )
 
 
-def build_fp_relu(fp_relu: FixedPointReLU) -> FPReLUModule:
-    return FPReLUModule()
+def build_fp_relu(fp_relu: FixedPointReLU, layer_id: str) -> FPReLUModule:
+    return FPReLUModule(
+        layer_id=layer_id,
+    )